Refactor navigate_to_child_programme function to utils and update test cases to use the new utility
